added licenses and separated commands to their own file
